requirements.txt
tensorflow==1.15.2
funcy
PyJWT==1.7.1
h5py==2.10.0
pillow
numpy
scipy
cython
matplotlib
scikit-image
opencv-python
imgaug
tensorflow-gpu==1.15.2
keras==2.2.4
IPython
sklearn